feat: a lot of improvements - improved folder bar design - a ton of improvements to code readability and effectiveness
