+++ Moc/Views/LoginView.swift
@@ -71,42 +71,6 @@ struct LoginView: View {
     @State private var showLoadingSpinner = false
 
     @Environment(\.presentationMode) private var presentationMode
-    
-    #if os(macOS)
-
-    func generateQRCode(from string: String) -> NSImage {
-        let context = CIContext()
-        let filter = CIFilter.qrCodeGenerator()
-
-        filter.message = Data(string.utf8)
-
-        if let outputImage = filter.outputImage {
-            if let cgimg = context.createCGImage(outputImage, from: outputImage.extent) {
-                return NSImage(cgImage: cgimg, size: NSSize(width: 32, height: 32))
-            }
-        }
-        
-        return NSImage(systemSymbolName: "xmark.circle", accessibilityDescription: nil)!
-    }
-    
-    #elseif os(iOS)
-    
-    func generateQRCode(from string: String) -> UIImage {
-        let context = CIContext()
-        let filter = CIFilter.qrCodeGenerator()
-        
-        filter.message = Data(string.utf8)
-        
-        if let outputImage = filter.outputImage {
-            if let cgimg = context.createCGImage(outputImage, from: outputImage.extent) {
-                return UIImage(cgImage: cgimg)
-            }
-        }
-        
-        return UIImage(systemName: "xmark.circle")!
-    }
-    
-    #endif
 
     var body: some View {
         // swiftlint:disable multiple_closures_with_trailing_closure
@@ -230,20 +194,16 @@ struct LoginView: View {
                         .padding(.top)
                     // QR Code
                     #if os(macOS)
-                    Image(nsImage: generateQRCode(from: qrCodeLink))
-                        .resizable()
-                        .interpolation(.none)
-                        .scaledToFit()
-                        .frame(width: 150, height: 150)
-                        .clipShape(RoundedRectangle(cornerRadius: 20))
+                    let image = Image(nsImage: .generateQRCode(from: qrCodeLink))
                     #elseif os(iOS)
-                    Image(uiImage: generateQRCode(from: qrCodeLink))
+                    let image = Image(uiImage: .generateQRCode(from: qrCodeLink))
+                    #endif
+                    image
                         .resizable()
                         .interpolation(.none)
                         .scaledToFit()
                         .frame(width: 150, height: 150)
                         .clipShape(RoundedRectangle(cornerRadius: 20))
-                    #endif
                     VStack {
                         stepView(number: 1, text: "Open Telegram from your phone")
                         stepView(number: 2, text: "Go to **Settings** -> **Devices** -> **Connect device**.")
