+++ iOS/GenerateQRCode.swift
@@ -1,0 +1,25 @@
+//
+//  GenerateQRCode.swift
+//  Moc
+//
+//  Created by Егор Яковенко on 14.06.2022.
+//
+
+import UIKit
+
+extension UIImage {
+    static func generateQRCode(from string: String) -> UIImage {
+        let context = CIContext()
+        let filter = CIFilter.qrCodeGenerator()
+        
+        filter.message = Data(string.utf8)
+        
+        if let outputImage = filter.outputImage {
+            if let cgimg = context.createCGImage(outputImage, from: outputImage.extent) {
+                return UIImage(cgImage: cgimg)
+            }
+        }
+        
+        return UIImage(systemName: "xmark.circle")!
+    }
+}
